feat(inventario): unidad METRO + soporte de decimales en cantidad
Pedido del cliente: agregar metros como unidad de medida con la
lógica que conlleva (decimales al vender/comprar).

Cambios:

1) Lista de unidades (UNIDADES_OPCIONES):
   - /inventario/nuevo y /inventario/[id]/editar: sumamos "METRO"
     (después de UNIDAD) y "CM" (para repuestos chicos en cm). Quedan
     antes que KG/G/LT para que sean fáciles de encontrar.

2) Helper nuevo src/lib/inventario/unidades.ts:
   - aceptaDecimal(unidad): true para METRO/M/CM/KG/G/LT/L/ML.
     Discretas (UNIDAD/CAJA/BOLSA/PAQUETE/DOCENA/etc.) → false.
   - sufijoUnidad(unidad): UNIDAD→u, METRO→m, KG→kg, etc.
   - formatCantidad(qty, unidad): "459.60 m" para decimales,
     "12 u" para enteros. Single source para futuros displays.

3) ProductPickerModal (carrito de ventas):
   - parseInt → parseFloat: acepta decimales en cantidad.
     Antes 2.5 m se redondeaba a 2 silenciosamente.
   - Input cantidad: step="0.01" cuando aceptaDecimal(unidad),
     step="1" sino. Label muestra la unidad: "Cantidad (metro)".

4) /ventas/nueva: cantNum = parseFloat (antes parseInt). Permite
   armar la línea con cantidades fraccionarias antes de mandar al
   carrito.

No requiere migración DB: productos.stock_actual ya es numeric, no
había restricción a enteros. Compras ya usaba parseFloat + step="any"
desde antes.

Lo que NO está incluido (avisado antes):
- El bug del importer (Excel "204.830" → 204830 por separador de miles)
  sigue ahí; el stock histórico de productos a granel quedó inflado.
  Se aborda en commit aparte cuando el cliente lo pida.
- Display masivo: no recorrí todas las pantallas reemplazando "u" por
  unidad real. Si lo querés, aplico formatCantidad() en /inventario,
  /ventas, etc.

Type-check OK.

+++ src/app/ventas/nueva/page.tsx
@@ -375,7 +375,9 @@ export default function NuevaVentaPage() {
   const tipoCambioNum = 1;
 
   const prodSel     = productos.find((p) => p.id === lineaProdId);
-  const cantNum     = parseInt(lineaCant) || 0;
+  // parseFloat para soportar decimales en unidades continuas (metros, kg, lt).
+  // Para UNIDAD/CAJA/DOCENA el front usa step=1 desde el input → siempre entero.
+  const cantNum     = parseFloat(lineaCant) || 0;
   const precioInput = parseFloat(lineaPrecio) || 0;
   const precioGs    = precioInput;
